% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_txt.R
\name{ajuste_txt}
\alias{ajuste_txt}
\title{Ajuste nas colunas de texto}
\usage{
ajuste_txt(df)
}
\arguments{
\item{df}{Dataframe a ser ajustado}
}
\value{
Retorna o dataframe com os ajustes realizados
}
\description{
Ajuste nas colunas de texto
}
