% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faixa_etaria.R
\name{faixa_etaria}
\alias{faixa_etaria}
\title{Cria faixa etária}
\usage{
faixa_etaria(df)
}
\arguments{
\item{df}{nome do Data frame}
}
\value{
Retorna uma coluna chamada faixa_etaria com as idades de
10 a 19 anos, 20 a 29 anos, 30 a 59 anos e 60+
}
\description{
Cria faixa etária
}
