% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padroniza_variaveis.R
\name{padroniza_variaveis}
\alias{padroniza_variaveis}
\title{Padroniza as variáveis}
\usage{
padroniza_variaveis(df, nome_base)
}
\arguments{
\item{df}{Nome do dataframe}

\item{nome_base}{nome da base, exemplo: "SINAN", "SIM" ou "SIH"}
}
\value{
Retora os nomes das variáveis padronizadas de acordo com a lista namestand
}
\description{
Padroniza as variáveis
}
