% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padroniza_variaveis.R
\name{padroniza_variaveis}
\alias{padroniza_variaveis}
\title{Padroniza os nomes das variáveis}
\usage{
padroniza_variaveis(df, df2, nome_base)
}
\arguments{
\item{df}{Nome do dataframe}

\item{df2}{Nome da tabela com os nomes padronizados}

\item{nome_base}{Nome da base a ser padronizada, ex: "SIM", "SINAN" ou "SIH"}
}
\value{
Retorna o dataframe com as colunas padronizadas
}
\description{
Padroniza os nomes das variáveis
}
